fix(stream): align StreamRef endpoint terminal handling
